Sound pool support, and move zoomed styles to CSS
Added a "soundPool" property to the square objects that specifies a pool
name to look up to play a random sound from. "sound" reverts to playing
the specified sound.

Also changed the zoom in/out functionality to add/remove a new CSS class
name from the object, so the specific values get to live in the .css.
The goal is that all styling should live in the .css, and all
configuration of the behavior should live at the top of the .js.