Blitz text-fit support
Plug up the textfit() calls in the right places, and set "display:
table-cell" and a max-height on #question-text so it can be resized, and
a width so it'll center properly (wouldn't work for single-line values
otherwise).

max-height of #question-text and .answer should be updated as necessary
to stay without the necessary bounds for this page.