Add multi-VGC "wall" display
To try and help reduce the overhead, return to the original bank of TVs
for VGC, with the ability to display the descriptions for any of those
challenges, plus Feelin' Sad and Grab Bag, including images.

This single view should supersede vgc.html, vgc-hard.html, and
vgc-easy.html, but I'm leaving the others in for the moment until we're
sure this is good.